A small cleanup of the option menu, makes it more modular so adding new options will not be as messy as before.
It adds a windowed mouse for GLX, and with the -DXMESA compiler option
(not currently supported by the configure/makefile process) there is an
additional option for users of Mesa3d running the quake-gl client  which
will switch between windowed and fullscreen modes on the fly.

The windowed mouse option is now fixed for windows users and would also
work for software-x11 clients, but #ifdef X11 isn't working?
--
Eric Windisch