beginnings of IPv6 support by Pontus Lidman
(pali) i had to personally fix this so that it's optional, if anyone wants to enable this (it's not ready yet, it's just a star) you have to uncomment the line\
for QW_NET_SRC in qw_client/Makefile.in and qw_server/Makefile.in, you also have to uncomment LINUX_IPV6 definition in qw_common/net.h

here's an excerpt of Pontus's explaination:
I've started to work on IPv6 support. I guess it's not a top priority, as
the intersections of quake players and people with 6bone access is pretty
small, but at least I'm one of them ;)

If anyone else is interested in working on IPv6 support, or have
opinions/advice on how to go about it, I'd like to get in touch.