uquake has the beginning of netchan support, netchan.message works fine now, the rest I don't know how much of it I can do without breaking the protocol's compatibility.
client_state_t has moved back to client.h in qw_common and uquake.  Seems
like that's a step backward, but the way it was being used while common
was far worse!  This required massive reworking of headers and such.

Speaking of using cl (the global client_state_t) badly, the sound code did
exactly that.  Mercury should be shot for not fixing that when he did the
sound_lib.a stuff.  The fix illustrates what we need to start doing to the
code to make modularization possible.  I'll be sending a message to the
list about this shortly..