Under a rare case in UQ (that you are playing with the QW HUD in a co-op game), the color bar that was supposed to be the backdrop for your frag counts would be drawn as if the statusbar was centered (which is never the case if you're using the HUD), fixed.
I have more sbar.c fixes to do but I need a working UQ to be able to test
them out first.  We're _almost_ there...  We all owe Deek and taniwha big
for the progress made tonight I think---great job guys.  But save SOME of
the merge for the rest of us willya?  ;>