SKYBOXES!
Relevant cvar to change the name of the skybox is r_skyname which defaults
to "sky".  If you don't have the files you'll get an ugly red background
in place of the sky.  I'll fix that shortly.

Sorry software people, this is just the Q2 code I'm turning on which doesn
not work for software.  The code is pretty simple though and most of it's
written already in gl_warp.c---just move that to r_sky and use the pcx's
instead of the tga's..