I decided I liked having a smaller console so I cleaned up the debugging code for scr_consize a bit.  I also made it work in software targets.  If you set scr_consize outside the range of .2 to .9 it now assumes you're an idiot and ignores your setting, using the appropriate boundary value.
A note about gl_conalpha for GL targets.  It now works up to a consize of
.9 rather than working just upto 2/3 of the screen.  Works just as it did
before.  The only reason it matters is that the boundary for scr_consize
does not apply to the decision of whether or not to use alpha--only to
how big to make the console.

That means if you set scr_consize to .91 or something it won't be bigger
but you'll lose alpha.  Don't even try to report that as a bug---I will
tell you simply Don't Do That(TM).