This SHOULD allow Mega2K to work whenever it gets released.  View offsets are done (sorry if this steps on your toes with the view.c merge Deek) and I'm almost positive flymode will now work.  Even though view offset is done, it won't.
The reason for this is that cl.stats[STAT_FLYMODE] is pretty much going to
ALWAYS be 0 on a standard server.  Since 0 tells us that we're not flying,
this is fine.  cl.stats[STAT_VIEWHEIGHT] is also going to be 0, but it
should be 22 for normal views.  I could always assume this value is an
offset from 22, but that just seems lame to me.  I'll either do it anyway
or we'll have to find a good opportunity in the connect cycle to set the
cl.qfserver qboolean to true.

I'm thinking about using an info key value for this, but we'd be better
served I think by coordinating with QSG to up the protocol version across
all engines.