Added cvars to configure (or disable) the server-side time cheat protection.
sv_timekick: This cvar controls the number of times a player has to
be caught "cheating" before they get kicked. sv_timekick shows up in
serverinfo if changed from the default. If sv_timekick < 1, timecheat
detection is disabled. Default is 3.

sv_timekick_fuzz: This cvar affects how strict the protection is. The
higher the number, the more "fuzz" is applied, and the less strict the
detection code is. Raise this if your players are being kicked for packet
loss and lag. Default is 10, giving a fuzz factor of about 1 percent.

sv_timekick_interval: This cvar controls how often, in seconds, the time
tally is counted. Lowering this value increases the chance of false
positives, but helps to minimize the amount of damage a timecheating
player can cause. Default is 30 seconds.