Here's the fireball code I promised, implemented from an email which told me pretty much how.  It's a meshing of the GL flashblend bubble and the particle engine.  More emphasis on dlights though.
This code should be very similar to an effect that was in QL.  It was
described to me in very explicit detail in an email.  It's probably based
on the QL source but is a legitimate clean-room implementation.  The email
described how to make the flame trail found in newer QL too but I've not
decided whether or not to code that yet - I think I can do better instead
by adding smoke.